+++ docs/changelog/changelog.rst
@@ -48,6 +48,7 @@ v0.7.11__develop
 * Fix logic for the 'auto' bin selection of `build_continuous_partition_object`
 * Add missing jinja2 dependency
 * Fix an issue with inconsistent availability of strict_min and strict_max options on expect_column_values_to_be_between
+* Fix an issue where expectation suite evaluation_parameters could be overriden by values during validate operation
 
 
 v0.7.10